+++ models/index.js
@@ -2,9 +2,9 @@ import mongoose from "mongoose";
 
 const { Schema, model } = mongoose;
 
-const userSchema = new Schema({
-  username: String,
-  password: String,
+export const userSchema = new Schema({
+  phoneNumber: String,
+  joinedOn: Date,
 });
 
 export const User = model("User", userSchema);
@@ -32,7 +32,7 @@ export const Message = model("Message", messageSchema);
 export const PERMITTED_AMIGO_STATUSES = ["lost", "found", "reunited"];
 
 // lost can have events with status (sighted) or even (recovered) before finally reunited
-// found just goes straight to reunited 
+// found just goes straight to reunited
 export const PERMITTED_STATUS_EVENT_STATUSES = ["sighted", "recovered"].concat(
   PERMITTED_AMIGO_STATUSES
 );
@@ -61,6 +61,7 @@ const amigoSchema = new Schema({
   owner_aware: Boolean, // can figure out later what the implications are but important to distinguish when the owner is *actually* taking care of the animal
   last_status_event: Object,
   last_updated_at: Date,
+  user: Object, // TODO this doesn't allow for changes, need to switch to Postgres
 });
 
 export const Amigo = model("Amigo", amigoSchema);
