+++ routes/amigos.js
@@ -1,7 +1,58 @@
 import express from "express";
-import { Amigo } from "../models/index.js";
+import { userSchema, User, Amigo } from "../models/index.js";
+// import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
 
 const amigosRouter = express.Router();
+amigosRouter.post("/users", async (req, res) => {
+  try {
+    const { phoneNumber } = req.body; // password
+    // const hashedPassword = await bcrypt.hash(password, 10);
+    const user = new User({ phoneNumber, joinedOn: new Date() }); // , password: hashedPassword
+    await user.save();
+    return res.status(201).json(user);
+  } catch (err) {
+    return next(err);
+  }
+});
+
+amigosRouter.patch("/users/:userId", async (req, res) => {
+  try {
+    const { userId } = req.params;
+    const { expoPushToken } = req.body;
+    const user = await User.findOne({ userId });
+    if (!user) {
+      return res.status(400).send({ error: "Could not find user." });
+    }
+    if (!expoPushToken) {
+      return next(new Error("need valid expoPushToken"));
+    }
+    user.expoPushToken = expoPushToken;
+    await user.save();
+    return next(user);
+  } catch (err) {
+    return next(err);
+  }
+});
+
+amigosRouter.post("/auth", async (req, res) => {
+  const { username, password } = req.body;
+  const user = await User.findOne({ username });
+  if (!user) {
+    return res.status(400).send({ error: "Could not find user." });
+  }
+
+  const passwordMatch = await bcrypt.compare(password, user.password);
+  if (!passwordMatch) {
+    return res.status(400).send({ error: "Invalid username or password." });
+  }
+
+  const token = jwt.sign(
+    { userId: user.id, username },
+    process.env.JWT_PRIVATE_KEY
+  );
+  return res.send(token);
+});
 
 amigosRouter.get("/amigos", async (req, res, next) => {
   try {
@@ -24,6 +75,15 @@ amigosRouter.get("/amigos/:amigo_id", async (req, res, next) => {
     } else {
       return res.json(amigo);
     }
+  } catch (err) {
+    return next(err);
+  }
+});
+
+amigosRouter.get("/users", async (req, res) => {
+  try {
+    const users = await User.find();
+    return res.json(users);
   } catch (err) {
     return next(err);
   }
